Add logic to defer play command response and respond to followup
