+++ src/commands/clear.rs
@@ -5,7 +5,7 @@ use serenity::{
     utils::Color,
 };
 
-use crate::utils::message::respond_to_command;
+use crate::utils::response::respond_to_command;
 
 pub async fn run(ctx: &Context, command: &ApplicationCommandInteraction) {
     let mut response_embed = CreateEmbed::default();
